+++ src/remix/codemods/root-v2.ts
@@ -1,4 +1,7 @@
+/* eslint-disable @typescript-eslint/no-unsafe-member-access */
 /* eslint-disable @typescript-eslint/no-unsafe-assignment */
+/* eslint-disable @typescript-eslint/no-unsafe-call */
+/* eslint-disable @typescript-eslint/no-unsafe-argument */
 
 import * as recast from 'recast';
 import * as path from 'path';
@@ -9,6 +12,8 @@ import type { ExportNamedDeclaration, Program } from '@babel/types';
 import { loadFile, writeFile } from 'magicast';
 
 import { ERROR_BOUNDARY_TEMPLATE_V2 } from '../templates';
+import { hasSentryContent } from '../utils';
+import { wrapAppWithSentry } from './root-common';
 
 export async function instrumentRootRouteV2(
   rootFileName: string,
@@ -63,15 +68,79 @@ export async function instrumentRootRouteV2(
 
     recast.visit(rootRouteAst.$ast, {
       visitExportDefaultDeclaration(path) {
-        // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
         const implementation = recast.parse(ERROR_BOUNDARY_TEMPLATE_V2).program
           .body[0];
 
         path.insertBefore(
-          // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
           recast.types.builders.exportDeclaration(false, implementation),
         );
 
+        this.traverse(path);
+      },
+    });
+    // If there is already a ErrorBoundary export, and it doesn't have Sentry content
+  } else if (!hasSentryContent(rootFileName, rootRouteAst.$code)) {
+    rootRouteAst.imports.$add({
+      from: '@sentry/remix',
+      imported: 'captureRemixErrorBoundaryError',
+      local: 'captureRemixErrorBoundaryError',
+    });
+
+    wrapAppWithSentry(rootRouteAst, rootFileName);
+
+    recast.visit(rootRouteAst.$ast, {
+      visitExportNamedDeclaration(path) {
+        // Find ErrorBoundary export
+        if (path.value.declaration?.id?.name === 'ErrorBoundary') {
+          const errorBoundaryExport = path.value.declaration;
+
+          let errorIdentifier;
+
+          // check if useRouteError is called
+          recast.visit(errorBoundaryExport, {
+            visitVariableDeclaration(path) {
+              const variableDeclaration = path.value.declarations[0];
+              const initializer = variableDeclaration.init;
+
+              if (
+                initializer.type === 'CallExpression' &&
+                initializer.callee.name === 'useRouteError'
+              ) {
+                errorIdentifier = variableDeclaration.id.name;
+              }
+
+              this.traverse(path);
+            },
+          });
+
+          // We don't have an errorIdentifier, which means useRouteError is not called / imported
+          // We need to add it and capture the error
+          if (!errorIdentifier) {
+            rootRouteAst.imports.$add({
+              from: '@remix-run/react',
+              imported: 'useRouteError',
+              local: 'useRouteError',
+            });
+
+            const useRouteErrorCall = recast.parse(
+              `const error = useRouteError();`,
+            ).program.body[0];
+
+            // Insert at the top of ErrorBoundary body
+            errorBoundaryExport.body.body.splice(0, 0, useRouteErrorCall);
+          }
+
+          const captureErrorCall = recast.parse(
+            `captureRemixErrorBoundaryError(error);`,
+          ).program.body[0];
+
+          // Insert just before the the fallback page is returned
+          errorBoundaryExport.body.body.splice(
+            errorBoundaryExport.body.body.length - 1,
+            0,
+            captureErrorCall,
+          );
+        }
         this.traverse(path);
       },
     });